requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
mangum
python-jose[cryptography]
passlib[bcrypt]
email-validator
typing_extensions==4.10.0
gitpython